calculations added
calculations added
