+++ EXPO20-main/EXPO20-main/main/src/Calculations.h
@@ -1,0 +1,3 @@
+int checkCurrentAmount(double currentAmount);
+
+int whichState(double currentAmount, double wantedAmount, double minimumCE);
