+++ EXPO20-main/EXPO20-main/main/src/main.cpp
@@ -2,6 +2,7 @@
 #include <Arduino.h>
 #include <Servo.h>
 #include <AccelStepper.h>
+#include <Calculations.h> 
 using namespace std;
 
 //Servo's
@@ -61,26 +62,7 @@ void switchSpeed(int theState){
     
 }
 //Alles van weight-------------------------------------------------|
-int checkCurrentAmount(double currentAmount){
-  // for now it is +=10
-  // later it will check the weight
-  currentAmount += 10;
-  return currentAmount;
-}
-int whichState(double currentAmount, double wantedAmount, double minimumCE){
-  if (currentAmount <= 0.8*wantedAmount){//normal speed under risk amount
-    return 10;
-  }
-  else if (currentAmount >= 0.8*wantedAmount){//slower speed above risk amount
-    return 20;
-  }
-  else if (currentAmount >= minimumCE){//stop when in CE range
-    return 30;
-  }
-  else{
-    return 0;
-    }
-}
+
 
 void setup() {
    //Arm void setup----------------------------------------------------------|
